Add styles to blog post
Add conditional rendering to poetry description so that it doesn't
show the excerpt. For instance, a Haiku wouldn't have been revealed
in the excerpt itself. Also added a link to go back to all writings
from within the blog post for easy navigation. Finally, changed the
layout of blog roll cards so that images are displayed on the side
with less prominence, particularly since the images are merely
thumbnails for decoration.