Fix styling on Card, Footer  and Layout component
Layout component now subtracts the height of the menu as well since
it is lowered by the menu's height to show the content.

Card component on hover is moved to the base container, so that
hovering on the image will also trigger the hover effect.

Footer height is set at the background, and not the paragraph.